Edited folder structure
