+++ config/config.py
@@ -1,3 +1,4 @@
+""" config/config"""
 import os
 basedir = os.path.abspath(os.path.dirname(__file__))
